pyproject.toml
[tool.poetry]
name = "synonym-api"
version = "0.1.0"
description = "An API for searching word synonyms. Only available in pt-BR."
authors = ["Igor Xavier <igoorlimaa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
Django = "^3.2"
djangorestframework = "^3.12.4"
beautifulsoup4 = "^4.9.3"
django-environ = "^0.4.5"
django-cors-headers = "^3.7.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
